+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Nginx sidecar to expose create archives
 - [#9] Collect volume information from prometheus
 - Regularly sync archives with cluster state to avoid finalizers
+- [#15] Garbage-collect old support archives
 
 ## [v0.2.0] - 2025-07-18
 ### Added